requirements.txt
# WebDriver and Selenium
selenium==4.15.2
webdriver-manager==4.0.1

# Testing Framework
pytest==7.4.3
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
pytest-timeout==2.2.0

# Reporting
allure-pytest==2.13.2
allure-python-commons==2.13.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and Utilities
python-logging-loki==0.3.1

# Data Management
jsonschema==4.20.0

# Code Quality
pytest-cov==4.1.0

# API Testing
requests==2.31.0

# Performance
pytest-benchmark==4.0.0